Add basic ADSR
MidiMapper attempts to reuse the voice that's currently
handling a key when (for example) ADSR release is in progress
and the key is pressed again.

ADSR isn't perfect in this case, still have to do with envelope
mixing, but it avoids two voices playing the same key for the time
being.